+++ controllers/admin/ProposedProcedureController.php
@@ -74,19 +74,26 @@ class ProposedProcedureController extends AdminBase
     /**
      * @param int $agendaItemId
      * @param int $comments
+     * @param int $onlypublic
      * @return string
      */
-    public function actionOds($agendaItemId = 0, $comments = 0)
+    public function actionOds($agendaItemId = 0, $comments = 0, $onlypublic = 0)
     {
         $this->consultation->preloadAllMotionData();
+
+        $comments   = (IntVal($comments) === 1);
+        $onlypublic = (IntVal($onlypublic) === 1);
 
         $filename = 'proposed-procedure';
         if ($agendaItemId) {
             $agendaItem      = $this->consultation->getAgendaItem($agendaItemId);
             $filename        .= '-' . trim($agendaItem->getShownCode(true), "\t\n\r\0\x0b.");
-            $proposalFactory = new Factory($this->consultation, true, $agendaItem);
+            $proposalFactory = new Factory($this->consultation, !$onlypublic, $agendaItem);
         } else {
-            $proposalFactory = new Factory($this->consultation, true);
+            $proposalFactory = new Factory($this->consultation, !$onlypublic);
+        }
+        if ($onlypublic) {
+            $filename .= '-public';
         }
         $filename .= '.ods';
 
@@ -97,7 +104,7 @@ class ProposedProcedureController extends AdminBase
 
         return $this->renderPartial('ods', [
             'proposedAgenda' => $proposalFactory->create(),
-            'comments'       => (IntVal($comments) === 1),
+            'comments'       => $comments,
         ]);
     }