+++ web/typescript/backend/ProposedProcedureExport.ts
@@ -5,9 +5,11 @@ export class ProposedProcedureExport {
 
     private recalcLinks() {
         let withdrawn = (this.$widget.find("input[name=comments]").prop("checked") ? 1 : 0);
+        let onlyPublic = (this.$widget.find("input[name=onlypublic]").prop("checked") ? 1 : 0);
         this.$widget.find("a.odsLink").each((ev, el) => {
             let link = $(el).data("href-tpl");
             link = link.replace("COMMENTS", withdrawn);
+            link = link.replace("ONLYPUBLIC", onlyPublic);
             $(el).attr("href", link);
         });
     }